WebSeamRepair/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
pillow

